refactor: remove API_ENDPOINTS from .env.example (#423)
This easily becomes out of date. When you start work on this project for the first time, copy .env.example and run the app it will be missing several endpoints. I don't believe we have a use case currently for overridding API endopints at run time, but I've left the ability to do that in anyway.